Move Construct & DeCon to Operation dir
