+++ Tunny/Component/Operation/ConstructFishAttribute.cs
@@ -8,7 +8,7 @@ using Grasshopper.Kernel.Parameters;
 using Tunny.Component.Params;
 using Tunny.Resources;
 
-namespace Tunny.Component.Util
+namespace Tunny.Component.Operation
 {
     public class ConstructFishAttribute : GH_Component, IGH_VariableParameterComponent
     {
@@ -23,7 +23,7 @@ namespace Tunny.Component.Util
         public ConstructFishAttribute()
           : base("Construct Fish Attribute", "ConstrFA",
             "Construct Fish Attribute.",
-            "Tunny", "Util")
+            "Tunny", "Operation")
         {
         }
