+++ Tunny/Component/Operation/ConstructFishEgg.cs
@@ -9,7 +9,7 @@ using Tunny.Core.Util;
 using Tunny.Type;
 using Tunny.Util;
 
-namespace Tunny.Component.Util
+namespace Tunny.Component.Operation
 {
     public partial class ConstructFishEgg : GH_Component
     {
@@ -19,7 +19,7 @@ namespace Tunny.Component.Util
         public ConstructFishEgg()
           : base("Construct Fish Egg", "ConstrFEgg",
             "You can specify the initial individual that Tunny will run. Try your golden egg!",
-            "Tunny", "Util")
+            "Tunny", "Operation")
         {
         }
